Extending the CodeMetropolis GUI with the components of the Mapping file editor. There is a known issue! See below for details.
I have also created the class for the editor dialog and made some minor changes on the BuildableSettings and PropertyCollector classes.
Known issue: in CodeMetropolisGUI.java you have to change the mappingEditorCdfPath.getText() in the line 215, as written in the comment.
If you don't do that you will get the error message that the file you have just chosen does not exist. If you replace the parameter, it
will work just fine. (An empty dialog should pop up without occuring any kind of exception.)